requirements.txt
ollama
argparse
Pillow
